Identity: One class per file
