+++ vector/src/main/java/im/vector/riotx/features/discovery/change/SetIdentityServerAction.kt
@@ -13,18 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.riotx.features.discovery
 
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import im.vector.riotx.core.extensions.postLiveEvent
-import im.vector.riotx.core.utils.LiveEvent
-import javax.inject.Inject
+package im.vector.riotx.features.discovery.change
 
-class DiscoverySharedViewModel @Inject constructor() : ViewModel() {
-    var navigateEvent = MutableLiveData<LiveEvent<DiscoverySharedViewModelAction>>()
+import im.vector.riotx.core.platform.VectorViewModelAction
 
-    fun requestChangeToIdentityServer(serverUrl: String) {
-        navigateEvent.postLiveEvent(DiscoverySharedViewModelAction.ChangeIdentityServer(serverUrl))
-    }
+sealed class SetIdentityServerAction : VectorViewModelAction {
+    data class UpdateServerName(val url: String) : SetIdentityServerAction()
+    object DoChangeServerName : SetIdentityServerAction()
 }
